fix(monitor_viewer): add positive offsets
Previously the monitor_viewer only had negative offsets that were
applied to the monitors that were behind or above the main one. This now
also adds positive padding offsets for the monitors that are in front or
below the main one. Only the main one has no offset.
This seems be to be working fine but needs further testing with more
complicated setups.