feat: always set default values to `None`
- Always set default values to `None`.
- `check_changes` now merges both configs (the current one and the
loaded config) with the `DEFAULT_CONFIG`, which returns a new config
with the values of the config passed cloned when `Some` and if `None` it
uses the value from `DEFAULT_CONFIG` instead. This makes it so if the
user had some config set with the same value as default then on the UI
changed the value from default to something else and back again, the
`sanitize_value` function would turn it to `None` which would be
different than the loaded_config value, but since we merge both with the
`DEFAULT_CONFIG` it ends up noticing that it was the default value so it
doesn't flag the config as dirty.
- In the future we might want to make this a setting or even change the
behavior since we might want the config to be marked dirty so that the
user can save it and change all default values to `None` to simplify
their configs as much as possible.